check for existing links or folders

+++ setup_srv_salt.sh
@@ -11,10 +11,18 @@ fi
 pushd /srv/saltshaker
 git submodule init
 git submodule update
-ln -sv /srv/saltshaker/srv/salt /srv/salt
-ln -sv /srv/saltshaker/srv/salt-modules /srv/salt-modules
-ln -sv /srv/saltshaker/srv/pillar /srv/pillar
-ln -sv /srv/saltshaker/srv/reactor /srv/reactor
+if [ ! -e /srv/salt ]; then
+    ln -sv /srv/saltshaker/srv/salt /srv/salt
+fi
+if [ ! -e /srv/salt-modules ]; then
+    ln -sv /srv/saltshaker/srv/salt-modules /srv/salt-modules
+fi
+if [ ! -e /srv/pillar ]; then
+    ln -sv /srv/saltshaker/srv/pillar /srv/pillar
+fi
+if [ ! -e /srv/reactor ]; then
+    ln -sv /srv/saltshaker/srv/reactor /srv/reactor
+fi
 chown -R salt:salt /etc/salt
 salt-call saltutil.sync_all
 systemctl restart salt-master
